言語設定が日本語以外の場合は英語ラベルが表示されるように修正（メッセージボックスとフィードバックリンク）
