+++ src/js/locales/en/index.js
@@ -1,0 +1,5 @@
+export const en = {
+  "unreadMessages": "Unread Messages",
+  "noMessages": "Message",
+  "sendFeedback": "Send Feedback"
+}
