+++ src/js/locales/ja/index.js
@@ -1,0 +1,5 @@
+export const ja = {
+  "unreadMessages": "メッセージがあります",
+  "noMessages": "メッセージ",
+  "sendFeedback": "フィードバックを送る"
+}
